requirements_production.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==23.9.1
werkzeug==2.3.7
jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.2
python-telegram-bot==20.7
schedule==1.2.0
apscheduler==3.10.4
